tests/requirements.txt
google-generativeai
streamlit
pandas
matplotlib
seaborn
numpy